requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.47
psycopg2-binary==2.9.6
Werkzeug==2.2.3
gunicorn==20.1.0
gspread==5.7.2
google-auth==2.14.1
